Fixes Hyper-V agent port disconnect issue
On Windows / Hyper-V Server 2008 R2, when a switch port
have to be disconnected because the VM using it was
removed, DisconnectSwitchPort will fail, returning an
error code and a HyperVException is raised.

If the VM's VNIC has been removed, disconnecting the
switch port is no longer necessary and the it will
proceed to remove the switch port.

This issue is not present in Windows / Hyper-V Server 2012,
as the switch ports are cleaned automatically.

Closes-Bug: #1374108

Change-Id: I729753c15c51c86655651b132fcf8eab78884eb5